requirements.txt
txt
fastapi
uvicorn
llama-cpp-python
pydantic
